refactor(core): switch signal listener id to usize, drop unused re-exports and dead code
- Signal::subscribe/unsubscribe now use usize instead of u64 for the
  subscription id (internal API change; signature is a breaking pub
  change so version moves to 0.24.0 per the documented minor-bump policy
  for breaking API work). The usize sentinel usize::MAX replaces
  u64::MAX for the stale-handle case. Listener storage and the
  removed-listener merge-back path share a new ListenerEntry type alias
  to keep Vec type expressions short.
- Drop the unused set_inactive method (and its trait declaration) from
  AnySignalInner: no call site remains after the slab was simplified in
  earlier PRs. Removing the trait method lets us stay at zero workspace
  warnings.
- Remove three unused super::* re-exports in engine/tests/{lighting,
  physics,raytracing}/mod.rs and two pub(crate) re-exports of empty
  r#impl modules (macros/src/html/mod.rs and example/src/page/raytrace/
  hook/mod.rs) that produced unused_imports warnings. The underlying
  fn.rs files still use euv_engine::* directly and keep the default
  constructor in scope, so no behaviour change.

+++ core/src/renderer/render/fn.rs
@@ -28,7 +28,7 @@ where
     write(element, &initial_value);
     let euv_id: usize = element.ensure_euv_id();
     let element_clone: Element = element.clone();
-    let subscription_id: u64 = signal.subscribe(move || {
+    let subscription_id: usize = signal.subscribe(move || {
         // No `is_connected` guard: the binding is torn down synchronously
         // by `cleanup_subtree` (via `take_binding_cleanups`) before the
         // element leaves the DOM, so a stale fire is impossible — and for
